Разбивает большие модули на мелкие
