+++ index.html
@@ -236,6 +236,8 @@
     </template>
 
   <script src="./js/functions.js"></script>
+  <script src="./vendor/pristine/pristine.min.js"></script>
+  <script src="./vendor/nouislider/nouislider.js"></script>
   <script src="./js/main.js" type="module"></script>
 </body>
 </html>
